[BE] UPDATE : #S08P12D104-370 환불 데이터 계산할 때 TypeCast(형변환) 오류 해결
Former-commit-id: 685471227dc23a5d09971cdb35fe926c0a407dd5

+++ backend/bpservice/src/main/java/kr/co/bpservice/service/kiosk/KBrollyReturnService.java
@@ -179,8 +179,8 @@ public class KBrollyReturnService {
         Price priceInfo = priceRepository.getPrice();
         Integer depositeMoney = priceInfo.getDepositeMoney();   // 보증금
         Map<String, ?> refundInfo = brollyPayLogRepository.findPayLogForRefund(brolly, depositeMoney, priceInfo.getMoney());
-        Integer period = (Integer) refundInfo.get("period");    // 이용기간
-        Integer price = (Integer) refundInfo.get("price");      // 이용금액
+        Integer period = ((Long) refundInfo.get("period")).intValue();    // 이용기간
+        Integer price = ((Long) refundInfo.get("price")).intValue();      // 이용금액
         Integer refundMoney = depositeMoney - price;            // 환불금액
